+++ src/main/java/contest/collectingbox/module/collectingbox/domain/repository/CollectingBoxRepositoryCustom.java
@@ -1,8 +1,7 @@
-package contest.collectingbox.module.collectingbox.domain;
+package contest.collectingbox.module.collectingbox.domain.repository;
 
 import contest.collectingbox.module.collectingbox.dto.CollectingBoxDetailResponse;
 
 public interface CollectingBoxRepositoryCustom {
     CollectingBoxDetailResponse findDetailById(Long id);
-
 }
